logic exported to spearated file for better readibility
